+++ src/utils/index.js
@@ -5,4 +5,6 @@ export getTotalByColumns from './getTotalByColumns';
 export colSpanCells from './colSpanCells';
 export toggleClass from './toggleClass';
 export flattenData from './flattenData';
+export translateDOMPositionXY from './translateDOMPositionXY';
+export isRTL from './isRTL';
 export * from './requestAnimationTimeout';